Fix #35: move methods to header file
The Arduino IDE can't optionally compile files based on library
presence, so we can't move methods to files if they need Catena .h
files to compile.

This is a quick hack; the right thing is to either integrate these
methods into the non-Catena class (preferred), or make a separate
library.